Added Download and Code Refactoring
