+++ pubspec.yaml
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.5.2+1
+version: 2.1.0+3
 
 environment:
   sdk: '>=2.18.2 <3.0.0'
@@ -54,6 +54,7 @@ dependencies:
   introduction_screen: ^3.0.2
   path_provider: ^2.0.11
   permission_handler: ^10.2.0
+  url_launcher: ^6.1.6
 
 dev_dependencies:
   flutter_launcher_icons: ^0.10.0
